update: 添加问题“2839.判断通过操作能否让字符串相等I”的代码和题解 (#1472) 2839: AC.cpp (#1471) - AC,100.00%,100.00%
docs: VsCode ssh remote下 Git 行内变动标记不显示
docs: refactor - 类似话题放到一起

+++ Codes/2946-matrix-similarity-after-cyclic-shifts.rs
@@ -2,15 +2,14 @@
  * @Author: LetMeFly
  * @Date: 2026-03-27 08:48:57
  * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2026-03-27 09:07:38
+ * @LastEditTime: 2026-03-27 09:08:38
  */
-// THIS CANNOT BE ACCEPTED
 impl Solution {
     pub fn are_similar(mat: Vec<Vec<i32>>, k: i32) -> bool {
         let m = mat[0].len();
         for row in mat.iter() {
             for j in 0..row.len() {
-                if (row[j] != row[(j + k) % m]) {
+                if (row[j] != row[(j + k as usize) % m]) {
                     return false;
                 }
             }
